pyproject.toml
[tool.poetry]
name = "fluss-server"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
balder = { version = "0.1.22", extras = ["subscription"] }
channels-redis = "^4.0.0"
django-health-check = "^3.17.0"
lok = "^0.2.0"
omegaconf = "^2.2.3"
rich = "^12.6.0"
django-taggit = "^3.0.0"
django-probes = "^1.7.0"
django-guardian = "^2.4.0"
psycopg2 = "^2.9.4"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
docker = "^6.0.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
